presets: Improve virt presets
The current presets are mostly fine, but the way they're set up
can lead to some confusing behaviors.

For example, on a freshly installed system, the units for the
virtqemud service are configured as

  $ systemctl list-unit-files | grep virtqemud
  virtqemud.service         enabled     enabled
  virtqemud-admin.socket    enabled     disabled
  virtqemud-ro.socket       enabled     disabled
  virtqemud.socket          enabled     disabled

Notice how the status for the sockets differs from the
corresponding presets, which is already a bit confusing but not
problematic from the functional point of view.

Now let's suppose that, after making modifications such as
switching from the modular daemons to the monolithic daemon, the
user wanted to return to the default configuration. Running

  $ for unit in virtqemud{,-ro,-admin}.socket \
                virtqemud.service; do
      sudo systemctl preset $unit;
    done

would do the trick, resulting in the same configuration shown
earlier.

However, things would be a lot different if one were to run the
ever so slightly different

  $ for unit in virtqemud.service \
                virtqemud{,-ro,-admin}.socket; do
      sudo systemctl preset $unit;
    done

Notice how this time the service file is listed before the
sockets, instead of the other way around. That's the only
difference between the two commands, but the resulting
configuration will now be

  $ systemctl list-unit-files | grep virtqemud
  virtqemud.service         enabled     enabled
  virtqemud-admin.socket    disabled    disabled
  virtqemud-ro.socket       disabled    disabled
  virtqemud.socket          disabled    disabled

which will not work correctly.

So the current setup is both a bit confusing and a bit fragile.
In order to improve the situation on both fronts, explicitly
mark as enabled by default those sockets that have, up until now,
been implicitly enabled by virtue of being listed as Also= in
the [Install] section of a service, such as virtqemud, that is
enabled by default.

https://issues.redhat.com/browse/RHEL-3231

Signed-off-by: Andrea Bolognani <abologna@redhat.com>