presets: Enable read-only/admin sockets for all virt services
Currently only the main socket gets enabled by default for
secondary drivers and support services, but in reality any
socket can be used to trigger socket activation so we want them
all to be enabled by default.

Note that libvirt commit

  commit 826931e95a38af8322f8ad069dc89117c6404a00
  Author: Andrea Bolognani <abologna@redhat.com>
  Date:   Mon Sep 25 17:24:34 2023 +0200

    systemd: Add Also between sockets

    This results in all sockets for a service being enabled when a
    single one of them is.

    The -tcp and -tls sockets are intentionally excluded, because
    enabling them should require explicit action on the
    administrator's part; moreover, disabling them should not result
    in the local sockets being disabled too.

    Signed-off-by: Andrea Bolognani <abologna@redhat.com>
    Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>

  v9.8.0-32-g826931e95a

which will be part of the upcoming libvirt 9.9.0 release (due in
November 2023) will cause secondary sockets for a service to be
enabled when the main socket is, so once again this change simply
aligns the Fedora presets with the behavior that users will
actually see on their systems.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>