DNF 5: set best=False, skip_if_unavailable=True
Fedora should override DNF 5's upstream defaults and set best=False and
skip_if_unavailable=True. These are the status-quo settings for DNF 4 on
Fedora.

The DNF team may propose changing these downstream settings in the
future, but for now, we should keep the default behavior of DNF 5
consistent with DNF 4 as we replace DNF 4 with DNF 5 in Rawhide.

Resolves https://bugzilla.redhat.com/show_bug.cgi?id=2219153 and
https://bugzilla.redhat.com/show_bug.cgi?id=2216205.