WSL: Install wsl-distribution.conf in /usr/lib
This isn't meant to be edited and exists to inform WSL about the
distribution configuration. Install the file in /usr/lib and symlink to
the location WSL expects.

Additionally, mark /etc/wsl.conf as config to not replace on update. It
allows users to configure host drive mounts, the default user, etc.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>