Tweak /etc/issue to avoid incorrect English
Currently the message displayed above the login prompt on virtual
consoles looks like this:

  Fedora Linux 41 (Server Edition)
  Kernel 6.12.10-200.fc41.x86_64 on an x86_64 (ttS0)

The "on an $arch" part results in correct English for some of
the existing architectures (e.g. x86_64 and aarch64) but not
for some others (e.g. ppc64le and riscv64).

Tweak the message so that correct English is produced regardless
of the architecture.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>