Enable vgauthd.service in addition to vmtoolsd.service
vmtoolsd.service has [Install] Also=vgauthd.service, so when the former
is enabled, the latter is too. But when systemctl preset-all is executed,
systemd sees that vgauthd.service defaults to disabled, and will disable
it. This causes a confusion when systemd does an automatic preset-all during
first boot, causing units to be marked as modified. To avoid the problem,
use the same preset for both units, so that it does not matter in which
order the preset operations are executed.

Also, remove a duplication.